Organization of the Hub and added pdf gen for evaluation
